+++ example/lib/loc_time_temp.g.dart
@@ -7,10 +7,10 @@ part of 'loc_time_temp.dart';
 // **************************************************************************
 
 LocTimeTemp _$LocTimeTempFromJson(Map<String, dynamic> json) => LocTimeTemp(
-      location: json['location'] as String,
-      time: DateTime.parse(json['time'] as String),
-      temp: (json['temp'] as num).toDouble(),
-    );
+  location: json['location'] as String,
+  time: DateTime.parse(json['time'] as String),
+  temp: (json['temp'] as num).toDouble(),
+);
 
 Map<String, dynamic> _$LocTimeTempToJson(LocTimeTemp instance) =>
     <String, dynamic>{
@@ -21,9 +21,10 @@ Map<String, dynamic> _$LocTimeTempToJson(LocTimeTemp instance) =>
 
 ListOfLocTimeTemps _$ListOfLocTimeTempsFromJson(Map<String, dynamic> json) =>
     ListOfLocTimeTemps(
-      locations: (json['locations'] as List<dynamic>)
-          .map((e) => LocTimeTemp.fromJson(e as Map<String, dynamic>))
-          .toList(),
+      locations:
+          (json['locations'] as List<dynamic>)
+              .map((e) => LocTimeTemp.fromJson(e as Map<String, dynamic>))
+              .toList(),
     );
 
 Map<String, dynamic> _$ListOfLocTimeTempsToJson(ListOfLocTimeTemps instance) =>
@@ -42,20 +43,20 @@ const _$LocTimeTempSchemaMap = <String, dynamic>{
     r'location': {
       r'type': r'string',
       r'description':
-          r'/// The location to get the temperature in (e.g. "New York, NY")'
+          r'/// The location to get the temperature in (e.g. "New York, NY")',
     },
     r'time': {
       r'type': r'string',
       r'format': r'date-time',
-      r'description': r'/// The time in the given time zone'
+      r'description': r'/// The time in the given time zone',
     },
     r'temp': {
       r'type': r'number',
-      r'description': r'/// The temperature in degrees Fahrenheit'
-    }
+      r'description': r'/// The temperature in degrees Fahrenheit',
+    },
   },
   r'required': [r'location', r'time', r'temp'],
-  r'$defs': {}
+  r'$defs': {},
 };
 
 const _$ListOfLocTimeTempsSchemaMap = <String, dynamic>{
@@ -66,9 +67,9 @@ const _$ListOfLocTimeTempsSchemaMap = <String, dynamic>{
       r'type': r'array',
       r'items': {r'$ref': r'#/$defs/LocTimeTemp'},
       r'description':
-          r'/// A list of locations with their current time and temperature'
-    }
+          r'/// A list of locations with their current time and temperature',
+    },
   },
   r'required': [r'locations'],
-  r'$defs': {}
+  r'$defs': {},
 };