+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   vs OpenAI. I got tired of checking for provider name vs. alias vs. handle. I
   just made it be the same thing it is in the provider table so that you can
   programmatically create a fully-qualified model string.
+- fixed a nasty fully-qualified model naming bug
+- Better docs!
 
 ## 0.8.2